feat: create image registry creds modal seperated
